@@ -2,7 +2,7 @@
 
 [package]
 name = "othello_rust"
-version = "1.10.0"
+version = "1.10.1"
 authors = ["Esgrove <esgrove@outlook.com>"]
 description = "A simple Othello CLI game implementation in Rust"
 edition = "2024"
@@ -10,12 +10,12 @@ license = "MIT"
 readme = "README.md"
 repository = "https://github.com/Esgrove/othellogame"
 build = "build.rs"
-rust-version = "1.85"
+rust-version = "1.89"
 
 [dependencies]
 anyhow = "1.0.99"
 base16ct = { version = "0.3.0", features = ["alloc"] }
-clap = { version = "4.5.47", features = ["derive"] }
+clap = { version = "4.5.48", features = ["derive"] }
 colored = "3.0.0"
 rand = "0.9.2"
 sha2 = "0.10.9"
@@ -28,11 +28,9 @@ path = "src/main.rs"
 incremental = true
 
 [profile.release]
-# Enabling LTO increases compile times by a lot,
-# Enable full LTO with "true" and setting codegen units to 1.
-lto = "thin"
+lto = true
 # https://doc.rust-lang.org/rustc/codegen-options/index.html#codegen-units
-codegen-units = 16
+codegen-units = 1
 
 [lints.rust]
 unsafe_code = "forbid"
